feat: refactor API key validation logic into a separate module and update references
