+++ .gitignore
@@ -22,3 +22,4 @@ Thumbs.db
 DOCUMENTATION_SUMMARY.md
 Agent.md
 PROJECT_EXPLANATION.md
+GEMINI.md
