+++ frontend/scripts/settings.js
@@ -10,57 +10,9 @@
  * - Prevents API key from appearing in URL by using preventDefault on form submission
  * - Stores API key only in localStorage (client-side storage)
  * - Provides clear error messages for invalid keys
- * 
- * @author Pablo Vega
- * @version 1.0.0
  */
 
-/**
- * Validates the format of an OpenAI API key.
- * 
- * @param {string} apiKey - The API key to validate
- * @returns {{isValid: boolean, error: string|null}} Result object
- */
-const validateApiKey = (apiKey) => {
-    const trimmedKey = apiKey.trim();
-
-    if (trimmedKey.length === 0) {
-        return {
-            isValid: false,
-            error: 'API key cannot be empty.',
-        };
-    }
-
-    if (trimmedKey.length < 45) {
-        return {
-            isValid: false,
-            error: 'API key is too short. OpenAI API keys are at least 45 characters.',
-        };
-    }
-
-    if (trimmedKey.length > 56) {
-        return {
-            isValid: false,
-            error: 'API key is too long. Please verify you copied the correct key.',
-        };
-    }
-
-    if (!trimmedKey.startsWith('sk-')) {
-        return {
-            isValid: false,
-            error: 'API key format is invalid. OpenAI API keys start with "sk-".',
-        };
-    }
-
-    if (!/^sk-[A-Za-z0-9_-]+$/.test(trimmedKey)) {
-        return {
-            isValid: false,
-            error: 'API key contains invalid characters. Only alphanumeric characters, hyphens, and underscores are allowed.',
-        };
-    }
-
-    return { isValid: true, error: null };
-};
+import { validateApiKey } from './validators.js';
 
 document.addEventListener('DOMContentLoaded', () => {
     /** @type {HTMLFormElement} */
